corrections style et certaines méthodes (Controllers)

+++ controller/ControllerRental.php
@@ -15,13 +15,12 @@ class ControllerRental extends Controller {
     }
 
     public function search_book() {
-        $profile = Controller::get_user_or_redirect();
+        $profile = $this->get_user_or_redirect();
         $books = Rental::get_rental_join_book_join_user_rentdate($profile->id);
         $title = "";
         $author = "";
         $date = "";
         $filter = "";
-
         if (isset($_POST["title"]) && isset($_POST["author"]) && isset($_POST["title"]) && isset($_POST["date"]) && isset($_POST["filtre"])) {
             $title = $_POST["title"];
             $author = $_POST["author"];
@@ -33,7 +32,7 @@ class ControllerRental extends Controller {
     }
 
     public function add_rental_in_basket() {// on recupere un user mais le champ id est vide
-        $user = Controller::get_user_or_redirect();
+        $user = $this->get_user_or_redirect();
         $users = User::get_user_by_username($user->username);
         $books = Rental::get_rental_join_book_join_user_rentdate($user->id);
         $id = 0;
@@ -43,17 +42,17 @@ class ControllerRental extends Controller {
         $members = User::get_all_user();
         $usertoAddRent = "";
 
-        if (isset($_POST["panierof"])) {
+        if (isset($_POST["idbook"]) && isset($_POST["panierof"])) {
             $usertoAddRent = User::get_user_by_id($_POST["panierof"]);
             $rent = Book::get_book_by_id($_POST["idbook"]);
             if (!Rental::cpt_basket_ok($usertoAddRent->id)) {
-                $msg = "Vous ne pouvez pas louer plus de 5 livres a la fois";
+                $msg = "Vous ne pouvez pas louer plus de 5 livres à la fois!";
             } else {
-                if ($user->id != $usertoAddRent->id) {
+                if ($user->id !== $usertoAddRent->id) {
                     $rental = new Rental($id, $usertoAddRent->id, $rent->id, NULL, NULL);
                     $rental->insert_book_without_rent();
                 } else {
-                    $rental = new Rental($id, $usertoAddRent->id, $rent->id, NULL, NULL);
+                    $rental = new Rental($id, $users->id, $rent->id, NULL, NULL);
                     $rental->insert_book_without_rent();
                 }
             }
@@ -65,50 +64,59 @@ class ControllerRental extends Controller {
     }
 
     public function add_rental_for_user_in_basket() {
-        $user = Controller::get_user_or_redirect();
-        $books = Book::get_all_books();
-        $msg = " ";
-        $usertoAddRent = "";
-        $members = User::get_all_user();
-        $datetime = date("Y-m-d H:i:s");
+        $user = $this->get_user_or_redirect();
+        if ($user->is_admin() || $user->is_manager()) {
+            $books = Book::get_all_books();
+            $msg = " ";
+            $usertoAddRent = "";
+            $members = User::get_all_user();
+            $datetime = date("Y-m-d H:i:s");
 
-        $usertoAddRent = "";
-
-        if (isset($_POST["panierof"])) {
-            $value = $_POST["panierof"];
-            $usertoAddRent = User::get_user_by_id($value);
-            $allrentofUser = Rental::get_this_rental_not_validate($usertoAddRent->id);
-            if ($user->id != $usertoAddRent->id) {
-                if (Rental::cpt_book_rented_ok($usertoAddRent->id)) {
-                    foreach ($allrentofUser as $rent) {
-                        $rent->update_rental_rentdate_for_user($usertoAddRent->id, $datetime);
+            if (isset($_POST["panierof"])) {
+                $value = $_POST["panierof"];
+                $usertoAddRent = User::get_user_by_id($value);
+                $allrentofUser = Rental::get_this_rental_not_validate($usertoAddRent->id);
+                if ($user->id != $usertoAddRent->id) {
+                    if (Rental::cpt_book_rented_ok($usertoAddRent->id)) {
+                        foreach ($allrentofUser as $rent) {
+                            $rent->update_rental_rentdate_for_user($usertoAddRent->id, $datetime);
+                        }
+                    } else {
+                        $msg = "Cet utilisateur a déjà 5 locations en cours!";
                     }
                 } else {
-                    $msg = "cet utilisateur a deja 5 location en cours";
+                    if (Rental::cpt_book_rented_ok($usertoAddRentr->id)) {
+                        foreach ($allrentofUser as $rent) {
+                            $rent->update_rental_rentdate($datetime);
+                        }
+                    } else {
+                        $msg = "Vous avez déjà 5 livres en location!";
+                    }
                 }
-            } else {
-                if (Rental::cpt_book_rented_ok($usertoAddRent->id)) {
+            }
+            if (isset($_POST["annuler"]) && isset($_POST["panierof"])) {
+                $usertoAddRent = User::get_user_by_id($_POST["panierof"]);
+                foreach ($allrentofUser as $rent) {
+                    $rent->delete_rental();
+                }
+            }
+            if (isset($_POST["solo"])) {
+                if (Rental::cpt_book_rented_ok($user->id)) {
                     foreach ($allrentofUser as $rent) {
                         $rent->update_rental_rentdate($datetime);
                     }
                 } else {
-                    $msg = "vous avez deja 5 livre en location";
+                    $msg = "Vous avez déjà 5 livres en location!";
                 }
             }
-        }
-        if (isset($_POST["annuler"]) && isset($_POST["panierof"])) {
-            $usertoAddRent = User::get_user_by_id($_POST["panierof"]);
-            foreach ($allrentofUser as $rent) {
-                $rent->delete_rental();
-            }
-        }
-
-        $getUserRental = $usertoAddRent->get_rental_join_book_join_user_by_user_not_rented();
-        (new View("book_manager"))->show(array("books" => $books, "profile" => $user, "UserRentals" => $getUserRental, "msg" => $msg, "members" => $members, "actualpanier" => $usertoAddRent));
+            $getUserRental = $usertoAddRent->get_rental_join_book_join_user_by_user_not_rented();
+            (new View("book_manager"))->show(array("books" => $books, "profile" => $user, "UserRentals" => $getUserRental, "msg" => $msg, "members" => $members, "actualpanier" => $usertoAddRent));
+        } else
+            $this->redirect();
     }
 
     public function get_basket() {
-        $user = Controller::get_user_or_redirect();
+        $user = $this->get_user_or_redirect();
         $books = Book::get_all_books();
         $msg = " ";
         $usertoAddRent = "";
@@ -126,7 +134,7 @@ class ControllerRental extends Controller {
     }
 
     public function del_one_rental_in_basket() {
-        $user = Controller::get_user_or_redirect();
+        $user = $this->get_user_or_redirect();
         $books = Book::get_all_books();
         $getUserRental = $user->get_rental_join_book_join_user_by_user();
         $msg = " ";
@@ -144,7 +152,7 @@ class ControllerRental extends Controller {
     }
 
     public function returns() {
-        $profile = Controller::get_user_or_redirect();
+        $profile = $this->get_user_or_redirect();
         $books = Rental::get_rental_join_book_join_user_rentdates($profile->id);
         $title = "";
         $author = "";
@@ -155,58 +163,67 @@ class ControllerRental extends Controller {
     }
 
     public function delete_rental() {
-        $profile = Controller::get_user_or_redirect();
-        $books = Rental::get_rental_join_book_join_user_rentdates($profile->id);
-        $title = "";
-        $author = "";
-        $date = "";
-        $filter = "";
-        if (isset($_POST["delrent"])) {
-            $delrent = Rental::get_rentals_by_id($_POST["delrent"]);
-            foreach ($delrent as $del) {
-                $del->delete_rental();
-            }
+        $profile = $this->get_user_or_redirect();
+        if ($profile->is_admin()) {
             $books = Rental::get_rental_join_book_join_user_rentdates($profile->id);
-        }
-        (new View("returns"))->show(array("profile" => $profile, "books" => $books, "title" => $title, "author" => $author, "date" => $date, "filter" => $filter));
+            $title = "";
+            $author = "";
+            $date = "";
+            $filter = "";
+            if (isset($_POST["delrent"])) {
+                $delrent = Rental::get_rentals_by_id($_POST["delrent"]);
+                foreach ($delrent as $del) {
+                    $del->delete_rental();
+                }
+                $books = Rental::get_rental_join_book_join_user_rentdates($profile->id);
+            }
+            (new View("returns"))->show(array("profile" => $profile, "books" => $books, "title" => $title, "author" => $author, "date" => $date, "filter" => $filter));
+        } else
+            $this->redirect();
     }
 
     public function update_rental_returndate() {
-        $profile = Controller::get_user_or_redirect();
-        $books = Rental::get_rental_join_book_join_user_rentdates($profile->id);
-        $title = "";
-        $author = "";
-        $date = "";
-        $filter = "";
-        $datetime = date("Y-m-d H:i:s");
-        if (isset($_POST["idbook"])) {
-            $returnRental = Rental::get_rentals_by_id($_POST["idbook"]);
-            foreach ($returnRental as $return) {
-                $return->update_rental_returndate($datetime);
-            }
+        $profile = $this->get_user_or_redirect();
+        if ($profile->is_admin() || $profile->is_manager()) {
             $books = Rental::get_rental_join_book_join_user_rentdates($profile->id);
-        }
+            $title = "";
+            $author = "";
+            $date = "";
+            $filter = "";
+            $datetime = date("Y-m-d H:i:s");
+            if (isset($_POST["idbook"])) {
+                $returnRental = Rental::get_rentals_by_id($_POST["idbook"]);
+                foreach ($returnRental as $return) {
+                    $return->update_rental_returndate($datetime);
+                }
+                $books = Rental::get_rental_join_book_join_user_rentdates($profile->id);
+            }
 
-        (new View("returns"))->show(array("profile" => $profile, "books" => $books, "title" => $title, "author" => $author, "date" => $date, "filter" => $filter));
+            (new View("returns"))->show(array("profile" => $profile, "books" => $books, "title" => $title, "author" => $author, "date" => $date, "filter" => $filter));
+        } else
+            $this->redirect();
     }
 
     public function cancel_rental_returndate() {
-        $profile = Controller::get_user_or_redirect();
-        $books = Rental::get_rental_join_book_join_user_rentdates($profile->id);
-        $title = "";
-        $author = "";
-        $date = "";
-        $filter = "";
-
-        if (isset($_POST["idcancel"])) {
-            $returnRental = Rental::get_rentals_by_id($_POST["idcancel"]);
-            foreach ($returnRental as $return) {
-                $return->cancel_rental_returndate();
-            }
+        $profile = $this->get_user_or_redirect();
+        if ($profile->is_admin() || $profile->is_manager()) {
             $books = Rental::get_rental_join_book_join_user_rentdates($profile->id);
-        }
+            $title = "";
+            $author = "";
+            $date = "";
+            $filter = "";
 
-        (new View("returns"))->show(array("profile" => $profile, "books" => $books, "title" => $title, "author" => $author, "date" => $date, "filter" => $filter));
+            if (isset($_POST["idcancel"])) {
+                $returnRental = Rental::get_rentals_by_id($_POST["idcancel"]);
+                foreach ($returnRental as $return) {
+                    $return->cancel_rental_returndate();
+                }
+                $books = Rental::get_rental_join_book_join_user_rentdates($profile->id);
+            }
+
+            (new View("returns"))->show(array("profile" => $profile, "books" => $books, "title" => $title, "author" => $author, "date" => $date, "filter" => $filter));
+        } else
+            $this->redirect ();
     }
 
 }
